Blueprint assignment table and constraints separation inited. #16
