+++ test/algebra/plonk/variable_base_scalar_mul.cpp
@@ -48,8 +48,8 @@ BOOST_AUTO_TEST_CASE(variable_base_scalar_mul_5_wires_test_case) {
 
 	using curve_type = algebra::curves::bls12<381>;
 	using TBlueprintField = typename curve_type::base_field_type;
-	constexpr std::size_t WiresAmount = 5;
-	using TArithmetization = zk::snark::plonk_constraint_system<TBlueprintField, WiresAmount>;
+	constexpr std::size_t WitnessColumns = 5;
+	using TArithmetization = zk::snark::plonk_constraint_system<TBlueprintField, WitnessColumns>;
 
 	zk::components::blueprint<TArithmetization> bp;
 
@@ -79,8 +79,8 @@ BOOST_AUTO_TEST_CASE(variable_base_scalar_mul_9_wires_test_case) {
 
 	using curve_type = algebra::curves::bls12<381>;
 	using TBlueprintField = typename curve_type::base_field_type;
-	constexpr std::size_t WiresAmount = 9;
-	using TArithmetization = zk::snark::plonk_constraint_system<TBlueprintField, WiresAmount>;
+	constexpr std::size_t WitnessColumns = 9;
+	using TArithmetization = zk::snark::plonk_constraint_system<TBlueprintField, WitnessColumns>;
 
 	zk::components::blueprint<TArithmetization> bp;
